+++ lib/ui/theme/theme.dart
@@ -9,6 +9,7 @@ class PostsTheme {
   static ThemeData from({
     @required ColorScheme colorScheme,
     @required IconThemeData iconTheme,
+    @required TextTheme mergeTextTheme,
   }) {
     return ThemeData.from(colorScheme: colorScheme).copyWith(
       appBarTheme: AppBarTheme(
@@ -29,6 +30,7 @@ class PostsTheme {
         focusColor: colorScheme.primary,
       ),
       buttonColor: colorScheme.primary,
+      textTheme: Typography.englishLike2018.copyWith().merge(mergeTextTheme),
     );
   }
 
@@ -48,7 +50,11 @@ class PostsTheme {
       onBackground: Color(0xFF646464),
     );
     final iconTheme = IconThemeData(color: Color(0xFF646464));
-    return from(colorScheme: scheme, iconTheme: iconTheme);
+    final mergeTextTheme = Typography.blackCupertino;
+    return from(
+        colorScheme: scheme,
+        iconTheme: iconTheme,
+        mergeTextTheme: mergeTextTheme);
   }
 
   static ThemeData get darkTheme {
@@ -67,6 +73,10 @@ class PostsTheme {
       onBackground: Color(0xFFFFFFFF).withOpacity(0.6),
     );
     final iconTheme = IconThemeData(color: Color(0xFFFFFFFF).withOpacity(0.6));
-    return from(colorScheme: scheme, iconTheme: iconTheme);
+    final mergeTextTheme = Typography.whiteCupertino;
+    return from(
+        colorScheme: scheme,
+        iconTheme: iconTheme,
+        mergeTextTheme: mergeTextTheme);
   }
 }